Get testing to work
Set debugging to false on FatdynetTest
Make that debugging conditional
Start and stop dynet properly
Streamline dynet tests
Move package of FatdynetTest
