+++ main/src/test/scala/org/clulab/processors/TestCluProcessor.scala
@@ -1,22 +1,12 @@
 package org.clulab.processors
 
-import org.clulab.FatdynetTest
-import org.clulab.dynet.ConstEmbeddingsGlove
-import org.clulab.dynet.Utils
-import org.clulab.processors.clu.CluProcessor
-import org.scalatest.{FlatSpec, Matchers}
-
 /**
   * Unit tests for CluProcessor
   * User: mihais
   * Date: 6/17/17
   */
 class TestCluProcessor extends FatdynetTest {
-  var proc = {
-    Utils.initializeDyNet()
-    new CluProcessor()
-  }
-  
+
   "CluProcessor" should "tokenize raw text correctly" in {
     val doc = proc.mkDocument("John Doe went to China. There, he visited Beijing.")
     doc.clear()
@@ -177,9 +167,4 @@ class TestCluProcessor extends FatdynetTest {
     deps.hasEdge(2, 9, "nmod") should be (true)
   }
   */
-
-  it should "stop" in {
-    proc = null
-    ConstEmbeddingsGlove.SINGLETON = null
-  }
 }
