+++ main/src/test/scala/org/clulab/processors/TestUniversalEnhancedDependencies.scala
@@ -1,16 +1,6 @@
 package org.clulab.processors
 
-import org.clulab.FatdynetTest
-import org.clulab.dynet.ConstEmbeddingsGlove
-import org.clulab.dynet.Utils
-import org.clulab.processors.clu.CluProcessor
-import org.scalatest.{FlatSpec, Matchers}
-
 class TestUniversalEnhancedDependencies extends FatdynetTest {
-    var proc = {
-    Utils.initializeDyNet()
-    new CluProcessor()
-  }
 
   "CluProcessor" should "parse some basic sentences correctly" in {
     var doc = proc.annotate("Ras1 is associated with cancer.")
@@ -124,10 +114,5 @@ class TestUniversalEnhancedDependencies extends FatdynetTest {
     doc.sentences.head.universalEnhancedDependencies.get.hasEdge(5, 4, "nsubj") should be(true)
     doc.sentences.head.universalEnhancedDependencies.get.hasEdge(5, 1, "dobj") should be(true)
     doc.sentences.head.universalEnhancedDependencies.get.hasEdge(5, 3, "dobj") should be(false)
-  }
-
-  it should "stop" in {
-    proc = null
-    ConstEmbeddingsGlove.SINGLETON = null
   }
 }
